fixing dbconnect logic, code architecture
